+++ index.d.ts
@@ -6,7 +6,7 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
-/* eslint-disable no-var, vars-on-top, camelcase, @typescript-eslint/triple-slash-reference */
+/* eslint-disable no-var, vars-on-top, @typescript-eslint/triple-slash-reference */
 
 /// <reference types="@v4fire/core"/>
 /// <reference path="./ts-definitions/playwright.d.ts"/>
@@ -20,10 +20,6 @@ declare module '*?raw' {
 declare var ssr: Nullable<{
 	document?: Document;
 }>;
-
-declare let __webpack_nonce__: CanUndef<string>;
-declare let __webpack_public_path__: CanUndef<string>;
-declare let __webpack_require__: (moduleId: string) => any;
 
 declare const BUILD_MODE: CanUndef<string>;
 
@@ -44,6 +40,13 @@ declare const BLOCK_NAMES: CanUndef<string[]>;
 declare const THEME: CanUndef<string>;
 declare const THEME_ATTRIBUTE: CanUndef<string>;
 declare const AVAILABLE_THEMES: CanUndef<string[]>;
+
+declare const DETECT_USER_PREFERENCES: CanUndef<
+	Dictionary<Dictionary<{
+		enabled: boolean;
+		aliases?: Dictionary<string>;
+	}>>
+>;
 
 declare const DS: CanUndef<DesignSystem>;
 declare const DS_COMPONENTS_MODS: CanUndef<{