+++ src/components/base/b-tree/b-tree.ts
@@ -62,7 +62,12 @@ class bTree extends iTreeProps implements iActiveItems, Foldable {
 
 	/** {@link iItems.items} */
 	set items(value: this['Items']) {
+		const oldValue = this.items;
 		this.field.set('itemsStore', normalizeItems.call(this, value));
+
+		if (this.isRelatedToSSR) {
+			this.syncItemsWatcher(this.items, oldValue);
+		}
 	}
 
 	/**